+++ cmd/server/main.go
@@ -11,10 +11,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
-
-	"github.com/livekit/protocol/auth"
 
 	"github.com/livekit/livekit-server/pkg/config"
 	"github.com/livekit/livekit-server/pkg/logger"
@@ -184,18 +181,18 @@ func startServer(c *cli.Context) error {
 	}
 
 	// require a key provider
-	keyProvider, err := createKeyProvider(conf)
-	if err != nil {
-		return err
-	}
-	logger.Infow("configured key provider", "numKeys", keyProvider.NumKeys())
+	//keyProvider, err := createKeyProvider(conf)
+	//if err != nil {
+	//	return err
+	//}
+	//logger.Infow("configured key provider", "numKeys", keyProvider.NumKeys())
 
 	currentNode, err := routing.NewLocalNode(conf)
 	if err != nil {
 		return err
 	}
 
-	server, err := service.InitializeServer(conf, keyProvider, currentNode, &routing.RandomSelector{})
+	server, err := service.InitializeServer(conf, currentNode, &routing.RandomSelector{})
 	if err != nil {
 		return err
 	}
@@ -210,31 +207,6 @@ func startServer(c *cli.Context) error {
 	}()
 
 	return server.Start()
-}
-
-func createKeyProvider(conf *config.Config) (auth.KeyProvider, error) {
-	// prefer keyfile if set
-	if conf.KeyFile != "" {
-		if st, err := os.Stat(conf.KeyFile); err != nil {
-			return nil, err
-		} else if st.Mode().Perm() != 0600 {
-			return nil, fmt.Errorf("key file must have permission set to 600")
-		}
-		f, err := os.Open(conf.KeyFile)
-		if err != nil {
-			return nil, err
-		}
-		defer func() {
-			_ = f.Close()
-		}()
-		return auth.NewFileBasedKeyProviderFromReader(f)
-	}
-
-	if len(conf.Keys) == 0 {
-		return nil, errors.New("one of key-file or keys must be provided in order to support a secure installation")
-	}
-
-	return auth.NewFileBasedKeyProviderFromMap(conf.Keys), nil
 }
 
 func getConfigString(c *cli.Context) (string, error) {
